+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - Support for autonomous transaction database-dedicated, autonomous exadata infrastructures, autonomous container databases and maintenance runs.
+- Support for `boot_volume_size_in_gbs` argument in the `oci_instance_configuration` resource 
 
 ## 3.28.1 (June 05, 2019)